Move cart related tests to payments app
